+++ package.json
@@ -51,6 +51,7 @@
     "fs-extra": "^4.0.2",
     "globby": "^7.1.1",
     "inquirer": "^4.0.1",
-    "mkdirp": "^0.5.1"
+    "mkdirp": "^0.5.1",
+    "pkg-dir": "^2.0.0"
   }
 }
