+++ src/WidgetDataInterface.ts
@@ -1,0 +1,4 @@
+export default interface WidgetDataInterface {
+	themeKey: string;
+	fileName: string;
+}
